+++ CHANGES.md
@@ -1,6 +1,13 @@
 <!-- markdownlint-configure-file { "no-duplicate-heading": { "siblings_only": true } } -->
 
 <!-- markdownlint-disable-next-line first-line-h1 -->
+## Unreleased
+
+### Added
+
+* vertigo-cli: `--watch-ignore-lists` parameter to ignore irrelevant files during watch (defaults to .gitignore) [#351]
+* vertigo-cli: `--global-ignores` to add custom wildcards to ignore during watch [#351]
+
 ## 0.7.0 - 2025-05-03
 
 ### Added